HADOOP-13327 Sync up with trunk; move implementation utils classes into fs.impl and inner classes to toplevel
